[metasrv] refactor: move sled db iniitializer util into metasrv/sled_store/
