+++ metasrv/src/sled_store/mod.rs
@@ -15,6 +15,17 @@
 //! sled_store implement a key-value like store backed by sled::Tree.
 //!
 //! It is used by raft for log and state machine storage.
+pub use db::get_sled_db;
+pub use db::init_sled_db;
+pub use db::init_temp_sled_db;
+pub use seq_num::SeqNum;
+pub use sled_serde::SledOrderedSerde;
+pub use sled_serde::SledSerde;
+pub use sled_tree::AsKeySpace;
+pub use sled_tree::SledTree;
+pub use sled_tree::SledValueToKey;
+
+pub mod db;
 pub mod seq_num;
 pub mod sled_key_space;
 pub mod sled_serde;
@@ -22,10 +33,3 @@ pub mod sled_tree;
 
 #[cfg(test)]
 mod sled_tree_test;
-
-pub use seq_num::SeqNum;
-pub use sled_serde::SledOrderedSerde;
-pub use sled_serde::SledSerde;
-pub use sled_tree::AsKeySpace;
-pub use sled_tree::SledTree;
-pub use sled_tree::SledValueToKey;
