+++ metasrv/src/tests/service.rs
@@ -25,10 +25,10 @@ use tempfile::TempDir;
 // use tracing_appender::non_blocking::WorkerGuard;
 use crate::api::FlightServer;
 use crate::configs;
-use crate::meta_service::raft_db::get_sled_db;
 use crate::meta_service::GetReq;
 use crate::meta_service::MetaNode;
 use crate::meta_service::MetaServiceClient;
+use crate::sled_store::get_sled_db;
 
 // Start one random service and get the session manager.
 #[tracing::instrument(level = "info")]
@@ -165,7 +165,7 @@ pub async fn assert_meta_connection(addr: &str) -> anyhow::Result<()> {
 macro_rules! init_meta_ut {
     () => {{
         let t = tempfile::tempdir().expect("create temp dir to sled db");
-        $crate::meta_service::raft_db::init_temp_sled_db(t);
+        $crate::sled_store::init_temp_sled_db(t);
 
         // common_tracing::init_tracing(&format!("ut-{}", name), "./_logs")
         common_tracing::init_default_ut_tracing();
